+++ server/src/index.ts
@@ -3,6 +3,7 @@ import { logger } from "./middleware/logger";
 import { errorHandler } from "./middleware/errorHandler";
 import { authRoutes } from "./routes/auth";
 import { artistRoutes } from "./routes/artists";
+import { gameRoutes } from "./routes/game";
 
 const app = new Elysia()
   .use(logger)
@@ -14,6 +15,7 @@ const app = new Elysia()
   }))
   .use(authRoutes)
   .use(artistRoutes)
+  .use(gameRoutes)
   .listen(3001);
 
 console.log(`🚀 Server running at http://${app.server?.hostname}:${app.server?.port}`);